scripts/requirements.txt
json
requests
pyyaml
